added userId to date and hourlyTime entities

+++ src/controllers/hourlyTime/deleteHourlyTime.ts
@@ -6,11 +6,18 @@ import { HourlyTime } from "../../entities/HourlyTime";
 export const deleteHourlyTime = (req: Request, res: Response) => {
     connectionToDB.then(async connection => {
         const hourlyTimeId = req.params.id;
+        const userId = req.body.userId;
         const hourlyTimeRepository = connection.getRepository(HourlyTime);
         let hourlyTime = await hourlyTimeRepository.findOne(hourlyTimeId);
 
         if (!hourlyTime) {
             unsuccessfulResponse(res, 404, 'time not found');
+            return
+        }
+
+        if (hourlyTime.userId !== userId) {
+            unsuccessfulResponse(res, 403, 'you are not authorized to delete this date');
+            return
         }
         
         await hourlyTimeRepository.remove(hourlyTime);
